separated parsing locally into doc_parser
